fix: arrumando rotas da aplicação (apenas isso)
